Crate peakbag

Find discrete returns in full waveform LiDAR data.

Why is this library called peakbag?

Peak bagging is when you try to summit a bunch of mountains, just to say you summited a bunch of mountains. While the practice of peak bagging can correlate with actual appreciation for the out-of-doors and an adventuresome spirit, a peak bagging attitude is neither necessary nor sufficient for a good time outside.

Some use "peak bagger" as a derisive term for someone who likes taking a selfie on top of a mountain more than just spending time outside.

This library finds peaks in waveforms, so peakbag seemed as good of a name as any.

Structs

Peak
A peak in the waveform data.

PeakDetector
Configurable struct for detecting peaks.

Functions

detect_peaks
Detects peaks in full waveform data.